refactor: rename gaia/models → gaia/gaia_ir
Better reflects that this package implements the Gaia IR specification.
Import path: from gaia.gaia_ir import KnowledgeNode, FactorNode
