+++ tests/gaia/libs/models/test_parameterization.py
@@ -5,7 +5,7 @@ from datetime import datetime, timezone
 import pytest
 from pydantic import ValidationError
 
-from gaia.models.parameterization import (
+from gaia.gaia_ir.parameterization import (
     CROMWELL_EPS,
     FactorParamRecord,
     ParameterizationSource,
